llm_server/pyproject.toml
[tool.poetry]
name = "llm-server"
version = "0.1.0"
description = ""
authors = ["Stuart Lynn <stuart.lynn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
websockets = "^12.0"
langchain = "^0.1.9"
langchain-community = "^0.0.24"
geopandas = "^0.14.3"
pandas = "^2.2.1"
langchain-openai = "^0.0.8"
ipython = "^8.22.1"
python-dotenv = "^1.0.1"
overpy = "^0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
